Move argument -> argument matcher transform
If we want to be able to use `InvocationShape` in verification methods
as well (as described in the previous commit message), it's likely
that this type will have to be able to accept arguments, instead of
argument matchers.

Therefore, extract that logic from `MethodCall` and move it into
`InvocationShape'.
