+++ src/Moq/InvocationShape.cs
@@ -40,7 +40,9 @@
 
 using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Linq;
+using System.Linq.Expressions;
 using System.Reflection;
 
 namespace Moq
@@ -52,6 +54,12 @@ namespace Moq
 	{
 		private readonly MethodInfo method;
 		private readonly IMatcher[] argumentMatchers;
+
+		public InvocationShape(MethodInfo method, IReadOnlyList<Expression> arguments)
+		{
+			this.method = method;
+			this.argumentMatchers = GetArgumentMatchers(arguments, method.GetParameters());
+		}
 
 		public InvocationShape(MethodInfo method, IMatcher[] argumentMatchers)
 		{
@@ -115,6 +123,21 @@ namespace Moq
 			}
 
 			return false;
+		}
+
+		private static IMatcher[] GetArgumentMatchers(IReadOnlyList<Expression> arguments, ParameterInfo[] parameters)
+		{
+			Debug.Assert(arguments != null);
+			Debug.Assert(parameters != null);
+			Debug.Assert(arguments.Count == parameters.Length);
+
+			var n = parameters.Length;
+			var argumentMatchers = new IMatcher[n];
+			for (int i = 0; i < n; ++i)
+			{
+				argumentMatchers[i] = MatcherFactory.CreateMatcher(arguments[i], parameters[i]);
+			}
+			return argumentMatchers;
 		}
 
 		private sealed class AssignmentCompatibilityTypeComparer : IEqualityComparer<Type>
